Use a span to hold the cursor position text
This fixes an issue where detaching and re-attaching the status bar
would cause the cursor position div not to be shown. The reason
why it wasn't shown is that it would get an inline display: block style
for some reason (overriding the display: inline-block), which breaks layout.
I'm guessing that this is a jQuery thing related to how it handles
hiding/showing DOM nodes.