Use specific icons for launch-mode indicator
The current icon is nebulous, and communicates very little to the reader
other than "something is different; check the tooltip to find out what."

This commit changes the launch-mode icon to use a safety jacket for safe
mode, and a microscope for dev mode. A beaker would have been preferred,
but it wasn't added until Octicons v3.0.0 (Atom's bundled with v2.1.2).