Fix exception when changing active pane item to a non editor
When the auto-save package is enabled and the whitespace package is set
to modify whitespace on the current line, the cursor position can change
synchronously as the result of changing the active pane item.
Previously, if we changed to a non-editor active pane item, the change
handlers in the status bar tiles would still assume an editor was active
and throw an exception.

This commit changes our handlers to always retrieve the last cursor from
the editor we actually subscribed to.