Add tooltip on filepath
Every couple of weeks, someone inside of Nuclide (Facebook IDE built on-top of Atom) support group asks how to copy filepath. Clicking on the filepath on the status bar isn't the first thing that comes to their mind.

Hopefully by adding a tooltip it's going to make it a small bit more discoverable :)

Test Plan:
![](http://g.recordit.co/VgDpLIGngT.gif)