Use tab length setting to calculate column number
Before now, the column number was a count of the number of characters to
the left of the cursor in the current row. This meant that hard tabs
were each counted as one character, even though they were displayed
according to the user’s tab-length setting.

This fix uses the user’s tab-length setting as the width of each tab
character when calculating the column number.